Add a Vagrantfile for local testing
Which mimics the four machines currently in `inventory`. This will be useful
for offline testing or working in parallel to other people.

I found that I had to write the inventory file dynamically because I
couldn't assign the `internal_ip` and `external_ip` variables when using
Vagrant's `groups` support. I think I've also mixed up the terminology for
"hosts" and "roles".

I'm using the PuppetLabs Ubuntu 14.04 box because it's pretty minimal and
supports both VirtualBox and VMware (which the official Ubuntu box doesn't).

There are some ordering issues to work out. It takes several runs on all
machines for them to complete successfully. This might account for the
non-numerical order of the original `inventory` file. We might be able to
help this by breaking each component out to it's own machine and ensuring
that databases happen first.