Remove `platform-add python` task
Nassim and I were unsure about adding this in the first place. I've now come
the conclusion that it's not terribly useful because it:

- depends on a Docker node which currently comes after the API node
- takes a long time and doesn't give any feedback about its progress
- pulls directly from GitHub master so may not be reproducible
- isn't used because we don't install the tsuru/tsuru-dashboard app

So I think we should remove it and do this step afterwards, along with other
platforms, perhaps still from Ansible, but not during the initial install.