Move handrolled ansible playbooks to new directory
https://www.pivotaltracker.com/story/show/91930038
We are moving the management of our playbooks to [librarian-ansible](https://github.com/bcoe/librarian-ansible) so
that we don't accidentally fork (or extend) our playbooks and we can
easily update them.

Some of our playbooks do not (yet) exist in
[ansible-galaxy](https://galaxy.ansible.com) so we are moving them to a
safe staging directory so librarian-ansible will copy the roles
directory.