Add librarian-ansible `Ansiblefile` manifest
Add librarian-ansible `Ansiblefile` which contains all the metadata
required to identify, find and install all our playbook dependencies.

We discussed inclusion of the `Ansiblefile.lock` file and decided to
include it so we can pin the dependencies.

We had an issue downloading the mongodb playbook using just the playbook
and version number that we worked around by using the github url, we are
not sure why this doesn't work but this is a sufficient fix for now.