Added docker registry storage backend options. By default the registry continues using the local file storage backend.
To be able to use S3 or GCS backend the option "storage_type"
should be changed, this can be done either by :
  * Uncommenting the option and set its value in the groupe_vars file.
  * Pass the option with -e on the command line
    (ex: "ansible-playbook ... -e 'storage_type=gcs'")

The storage credentials (S3/GCS keys) when needed are passed
using the  environment variable:
 AWS_ACCESS_KEY, AWS_SECRET_KEY, GCS_ACCESS_KEY, GCS_SECRET_KEY

Ex: running the playbook and setup GCS backend.
  # GCS_ACCESS_KEY=XXX GCS_SECRET_KEY=XXX ansible-playbook ... -e "storage_type=gcs"

Note that in case the instance on which the docker registry is running is bind to
an IAM role (for AWS) or a service account (for GCS) there is no need to pass the
credentials, the docker-registry servier collect them from the instance metadata, though
in case of GCS the option gcs_oauth2 should be set to 'true'.