Add ansible-vault support to encrypt secret data
We need to put the `login` and `password` data for `postgresql` into
this repository but I don't want to put it in clear text so we are
going to use
[ansible-vault](http://docs.ansible.com/playbooks_vault.html) to
encrypt the data. Our use of `ansible-vault` has a dependency on
version `1.9 or higher` of `ansible`

Taking this opportunity to encrypt the current tsuru credentials in
preparation for configuring credentials for postgresql.

Updated the documentation to show how to use `ansible-vault`.

Updated `Vagrantfile` so it uses `ansible.ask_vault_pass` to
unlock the password stored in your `ansible vault`
password.