Change internal ips to point at load balancer
These were the IPs of individual instances rather than an ELB and
`hipache_host_internal_lb` (which wasn't actually used) was only
pointing to the first of them. By changing it to the CNAME which points
at the ELB we're now using both of them.

We will address the 'correct' naming of the variables in a later story.