Split site.yml into platform-specific and common
As we are using the same ansible scripts to configure hosts on two different cloud providers, we are starting to need some platform-specific options passed to ansible roles. Specifically, we need the storage_type parameter for docker-registry to be set to 's3', 'gcs' or 'local' for AWS, GCE and Vagrant respectively.

- Renamed site.yml to common.yml
- Created site-aws.yml and site-gce.yml which include common.yml as their last step.
- Moved docker-registry host definitiion from common.yml into site-aws.yml and site-gce.yml
- Added storage configuration options to site-aws and site-gce
- Removed group_vars/docker-registry/storage_backend
- Added S3 access key and secret key to ansible vault
- Updated README.md to reflect changes.

Added gcs storage options to site-gce.yml

These variables are not yet in use, but adding them here for future reference.

Added s3_storage_path to site-aws for clarity

Just using the default path of "/".
Added purely for clarity in the S3 options