Replace curl commands with Ansible uri module
These shell-outs and curl commands weren't registering an error when the
admin user wasn't created and/or it was unable to login. It would write an
empty `~/.tsuru_token` file and a later task which adds a docker node would
hang indefinitely because it triggered an interactive login.

Fix this by using Ansible's uri module which checks status codes and allows
us to parse the JSON response without piping it through another command.