Use certifi for verifying GCE SSL certs
Because of epic yak shaving.

If you've installed Python from Homebrew, it compiles against a version of
OpenSSL that doesn't appear to validate `accounts.google.com`. We can't
reference Mac's list of CAs because they no longer use OpenSSL and are now
in Keychain. `curl-ca-bundle` did work in the past but was removed in
Homebrew/homebrew@ab926db. Everyone else appears to download or vendor their
own.

certifi seems like a reasonable compromise. It's maintained by the venerable
Kenneth Reitz and used by Requests. I did think about putting this in
`gce.py` itself, but I think it's better to not modify that file any further
and to make it obvious what we're doing.