Move ansible-playbook commands to Makefile
These are getting quite complex, so we should move them to something more
automated to reduce the chance of human error.

By doing so we can also check that the environment name is set.

I realise that this disadvantages people using Linux instead of Mac because
they are forced to use `vault_password.sh` instead of their own solutions
like reading from a file. It seems like a reasonable compromise though.