Render ssh.config based on environment name
`ssh_config(5)` doesn't support interpolation of environment variables, so
we have been updating this by hand for each independent environment. This is
a pain and will become worse when we need to do the same in Jenkins. So
render it based on a crude template and `gitignore` the resulting file.