Remove Vagrant
I initially added this when we had a single shared environment could work on
stories locally without affecting other people. Now that we have the ability
to create our own environments for each story, we're using dynamic
inventories and making more use of features that only exist in cloud
providers (registry object storage, load balancer appliances, etc), this
isn't so useful anymore and we haven't been maintaining it. So remove it :(