Address registry using DNS again
A regression was introduced in the dynamic inventory work (not a surprise
because it was very complex - alphagov/tsuru-ansible#41) which meant that we
went back to address the Docker registry by IP instead of hostname. This
causes problems when the IP of the registry machine is changed because
images are namespaced by the registry that they come from.

Go back to using DNS like we did in 42e3bfb1dc0bf1a7935917ff0d4e574ab29fe332
for the (then) static inventory.

I'm also removing the duplicate entry for `docker_registry_port`. We only
need the entry in globals, for both platforms.