Abstract difference in AWS/GCE tags
By using an intermediate variable. This doesn't allow us to de-dupe any
variables, but it does make the `hosts` entries much easier to read and mean
that we don't need to use regexps (which are error prone).

I've tested this with the following commands and verified that the lists of
hosts under each "play" remain the same:

    time make aws DEPLOY_ENV=dcarley ARGS=--list-hosts
    time make gce DEPLOY_ENV=dcarley ARGS=--list-hosts