Combine sslproxy and router roles
Combine the Nginx which does TLS termination and reverse-proxying for the
routers onto the same machines as the routers. This means that:

- the address of the upstream doesn't change like DNS records for ELBs do
- we're not subject to extra network latency/partitions for reverse proxying
- we have a simpler architecture

Hipache has been moved from `0.0.0.0:80` to `127.0.0.1:8080` so that it's
not accessible without going through Nginx and it doesn't clash with the
Nginx vhost that does HTTP->HTTPS redirection.

I've changed the format for passing `vars` to `roles` to make it clearer
which role is using which variables. The new `hipache_port` var is
referenced by vars passed to both of the roles.

This depends on alphagov/tsuru-terraform#57 being merged at the same time.