Update docker registry playbook
Currently we are using the interoperablity access method, however this
requires that each developer needs to generate and manage their own
`gcs_access_keys` and `gcs_secret_keys`.

We see this as an operational overhead that we don't want to bear with
in the long term so we have chosen to use a [service account](https://developers.google.com/identity/protocols/application-default-credentials) to share the
credentials across the team and make it easier for us to do continuous
delivery.

We need an [updated playbook](https://github.com/codingbunch/ansible-docker-registry/pull/3) with the google service account support in
order to do the above.