Move SSL proxy tasks and vars to separate files
So that they can be re-used by another host which needs to do the same
thing. I've renamed the one variable that we'll need to modify for each use
case from `hipache_port` to `upstream_port`.

I did try to include both the `pre_tasks` and the `role` within a single
file but I can't see any way to do this. The `vars_files` thing feels quite
hacky but is slightly better than duplicating code.