Nginx SSL reverse proxy in front of API
So that we can secure access to the API on all providers.

We don't currently have SSL on the API in GCE environments because the load
balancers which do SSL termination aren't currently supported by Terraform.
We do have ELBs terminating SSL in AWS, but the certificate upload process
is outside of Terraform.

This won't currently be used. I'm going to roll this out in several steps
because we don't currently have an SSL certificate for GCE and the Tsuru
client won't work on cert validation errors. These steps will follow:

- change the AWS load balancer to forward to TCP:443 instead of terminating
  SSL (but retain HTTP:8080 forwarding for the moment)
- add TCP:443 forwarding to the GCE load balancer

Then when we have the GCE certificate deployed to the right machines:

- remove all HTTP:8080 listeners and forwarders on load balancers
- change the API to only listen on `127.0.0.1:8080`

I was originally hoping to implement this using Tsuru's `use-tls` option
which allows you to pass a certificate and key for Tsuru to do SSL
termination. However this would have made the local `tsuru` and
`tsuru-admin` commands fail, which hit the API over loopback to do things
like register Docker nodes, due to SSL certificate hostname validation.

That could have been overcome by calling out to the internal load balancer
but that would have required waiting for enough health checks to pass for
the instances to be brought into service and it all seemed like unnecessary
complexity.