Run ansible-galaxy only when necessary
Currently we run the ansible-galaxy every time we run make,
even when it is not required if `requirements.yml` did not
change.

In this commit using a file as a flag to keep track of the
last run of `ansible-galaxy` task. `ansible-galaxy` will only
run if the task never run or `requirements.yml` is newer.