Use a ssh-wrapper instead of GIT_SSH_COMMAND
Older versions of git < 2.3 do not allow use the GIT_SSH_COMMAND variable
but only specify the ssh binary with GIT_SSH. Because that, in order to
specify the configuration file we need to create a wrapper.

In this commit we add a helper to create a wrapper and set it in the
endtoend test.