Disable nginx proxy buffering
Buffering is on by default on Nginx. It will cause it to wait for 1KB of output or close of connection before it sends reply back to original requestor. This affects situations where you want to stream response, e.g. when connecting to API server and adding platform or streaming logs. Disabling buffering makes Nginx behave more like ELB or google LBs in being more transparent and not interfering with the data flow from the apps.