Use random names for db instance on smoke tests
We generate a random name for the created db instance in the smoke
tests. This will allow us to run several smoke tests in parallel and
avoid any conflict.

Note: postgresql api does truncate the name of the db instance to
generate the assests in the postgresql db (db name, permissions, etc).
Because that, this code ensures that the first charaters are the most
variant.