Makefile does noise if it fails opening the vault
We retrieve the Tsuru credentials from the ansible vault, but if it
fails getting it, it will not display the error because we capture the
stderr from the `ansible-vault` command.

In this commit we make the `open_the_vault.sh` less verbose to keep
output clean, and we stop capturing the stderr when calling
`ansible-vault`. This way, if it fails, we can troubleshoot quickly.