Add tests for Tsuru API healthchecks
To test that all of the components (MongoDB, Router, Registry, etc) are
passing health checks by using this endpoint:

- http://docs.tsuru.io/en/master/reference/api.html#full-healthcheck-of-all-tsuru-components

Technically we don't need to check the output because the endpoint should
return a non-200 response if any of the components are not `WORKING`.
However for completeness I am checking that that we have at least three
lines of output so that the `each` loop gets called and that each line
contains `WORKING` in roughly the right place.

I've pulled some variables that I need into the outer `describe` block so
that I can use them across both contexts. I've also referenced
`@tsuru_api_host`, since we're setting it, rather than going back to
`RSpec.configuration.target_api_host` each time.