Use certifi CA bundle for end-to-end tests
Otherwise `open-uri` raises a certificate verify exception for the GlobalSign
certificate that we have on the API when running the tests from Mac OS X:

    Failures:

      1) TsuruEndToEnd tsuru API should pass healthchecks for all components
         Failure/Error: response = URI.parse("#{@tsuru_api_url}/healthcheck/?check=all").open()
         OpenSSL::SSL::SSLError:
           SSL_connect returned=1 errno=0 state=unknown state: certificate verify failed
         # ./spec/endtoend/endtoend_spec.rb:15:in `block (3 levels) in <top (required)>'
         # ./vendor/bundle/ruby/2.2.0/gems/rspec-retry-0.4.0/lib/rspec/retry.rb:43:in `block (3 levels) in apply'
         # ./vendor/bundle/ruby/2.2.0/gems/rspec-retry-0.4.0/lib/rspec/retry.rb:34:in `times'
         # ./vendor/bundle/ruby/2.2.0/gems/rspec-retry-0.4.0/lib/rspec/retry.rb:34:in `block (2 levels) in apply'

We could install the Ruby version of `certifi` certificate path, however it
seems to require monkey-patching to tell `open-uri` to always use that CA path.
Using the environment variable is easier and at least consistent with how we
run Ansible.