Split healthcheck assertions into separate tests
So that we can see which individual components have failed before gauging
the overall health from the response status code. We may want to disable
rspec `fail_fast` option in the future so that we can see all test failures.

I thought about enumerating a new `it` block for each components but we'd
have to do too much parsing of the output to produce reliable descriptions.