Copy script for adding datasources to Grafana
Access to Grafana API for adding a datasource usually requires
that you have an API key for your requests - However, in order
to get an API key, you need to log into the web interface.

The datasource.sh script works without using an API key by
navigating the login screen to retrieve a session cookie which
is then used in the subsequent request to add a datasource.

The script is a variation on the script produced by
@leehambley on a [github gist](https://gist.github.com/leehambley/9741431695da3787f6b3)