Specify team when creating new apps.
Our demo environment has identified a weakness in our current tsuru-ansible
scripts whereby we've never considered the case that the admin user may be a
member of more than one team. If this is the case than a `tsuru app-create` will
fail with an `Error: You belong to more than one team, choose one to be owner for this app.`

This commit adds the `-t` flag to the app-create commands to specify the `admin` team during
the post-install.yml ansible run.