Add a fake platform to do 31s delay while deploy
In order to implement a test on the timeout while deploying
applications, we need to simulate a delay > 30s while deploying, which
will trigger the heartbeat of the tsuru API.
The `tsuru_unit_agent` implements other heartbeat of 5s, so we need to
catch the calls to `tsuru_unit_agent`.

This is a non functional platform, it just does sleep 30 for every
`tsuru_unit_agent` execution.