Add logic to handle locked app in test after hook
Sometimes the smoke tests don't clean properly the apps, leaving them
running in the platform.

After investigating, we found that when running the test the application
can be potentially locked in the tsuru API server. For instance,because the
git push timeout or was disconnected. When this happens, the `after` hook
will fail cleaning up the application.

This code adds logic which will wait 5 seconds and retry if it fails
deleting it because it is locked, giving a change to finish the running
task. After 5 retries, it will force an unlock to delete it.