Set exit_status to != 0 if child process is killed
Solves the case of getting a `exit_status` == 0 if the child process
exits due an unhanded signal.

When calling a process with Popen3 the resulting value attribute
(`Process::Status`) converted to int with `to_i` can return the
exit code shifted 8 bit or the signal if it has been signalled.

Now the exit_status will be set to the signal number + 128 as most
of the shells return.