Use existing variable to select tsuru-api-0 host
The order of the hosts array isn't always consistent as demonstrated by:

    (ansible)➜  tsuru-ansible git:(master) ✗ ansible all -i ec2.py -m shell -l 'tag_Name_dcarley-tsuru-api*[0]' -a 'hostname'
    10.128.13.46 | success | rc=0 >>
    ip-10-128-13-46

    (ansible)➜  tsuru-ansible git:(master) ✗ ansible all -i ec2.py -m shell -l 'tag_Name_dcarley-tsuru-api-0' -a 'hostname'
    10.128.11.176 | success | rc=0 >>
    ip-10-128-11-176

This was causing the Docker "register node" task to occasionally hang
because it was delegated to the *other* API node, which hadn't executed the
tasks to login and create `~/.tsuru_token`, so `tsuru-admin` was waiting for
an interactive login.

We have to use `tsuru_api_host_name` rather than `tsuru_api_host` (as used
in `delegate_to`) because `hosts:` expects a tag/host matcher rather than an
IP address.