Close STDIN on some tsuru-admin commands
So that these tasks fail early with an error instead of waiting forever on
an interactive login when `~/.tsuru_token` is absent or invalid.

I did try changing these to use the `command` module in the hope that it
closed STDIN automatically, but it doesn't:

- http://docs.ansible.com/ansible/command_module.html

I also looked at a way to specify a timeout of how long to wait for a
command but it seems you need to use the `async` module to do this which
seems like too greater change:

- http://docs.ansible.com/ansible/playbooks_async.html

There are other `tsuru` and `tsuru-admin` commands in our codebase that I
haven't touched. It seems more important to fix these ones first because
they hang early in the deployment and prevent the core components from being
setup, whereas if the post-install hangs then you already have a partially
working environment and it should take less time to fix.