Retry deletion of the service if it is still bound
Tsuru application removal code happens asynchronously in the API server,
so the call to `tsuru app-remove` will return immediately despite it has
not been deleted and the services are bound.

Because that, if we try to delete the services immediately after the
app, it might result in a error: `Error: This service instance is bound to at least one app. Unbind them
before removing it`

To workaround this issue we add logic to retry after sleeping 1 second.
I assume that 25 retries with 1 second delay is enough.