Add self-referencing group to gce.py
This is a hack so that we can de-duplicate the variables in
`platform-aws.yml` and `platform-gce.yml` by referring to the `hostvars` of
other nodes in a consistent way, e.g.

    `hostvars[groups["{{ hosts_prefix }}-tsuru-api-0"][0]]`

This pattern is necessary on AWS because the `hostvars` dictionary is keyed
by the instance IP rather than it's `Name` tag and tags are not unique
across instances.

It might have been possible to do the same using jinja2 filters to select
entries with the appropriate attributes, but it looked to produce much more
complicated code.

I don't normally comment code, but this seems warranted since it's not at
all obvious why the code is doing that thing and we may want to find a
better solution in the future.