Don't use var in hosts pattern for first API node
The preceding commit moved this variable from `platform-{aws,gce}.yml` to
`group_vars/all/globals.yml`. This seems to prevent Ansible from
interpolating this variable correctly:

    PLAY [{{ tsuru_api_host_name }}] **********************************************
    skipping: no hosts matched

Fix this by changing to a hosts pattern that only matches the first API
node. This is in keeping with the spirit of 3ff22e4, but unfortunately
leaves us with some duplication.