post-install: run explicitly on api-0
Use `{{ hosts_prefix }}-tsuru-api-0` instead of `{{ hosts_prefix }}-tsuru-api*[0]` to indicate which host to run on. This is the same host name as used in tsuru_api.yml, which does some pre-requisite steps for post_install. This way we ensure we always run on the same host, as *[0] can sometimes point to different server.