Install all Tsuru packages from our PPA mirror
Install the Tsuru packages for the core components Hipache, Gandalf, and
API, from our own PPA which is a mirror of the upstream PPA. This allows us
to pin the version that we use for stability.

A limitation of PPAs whereby they only present the latest version meant that
if we have pinned the version number in our code then we would have been
unable to install when Tsuru release a new version.

We have to use `pre_tasks` to remove the upstream PPA which is no longer
installed by the playbooks when the `tsuru_repo` var is passed. This can be
removed once it has been deployed to all existing machines, so long as we
continue to pass `tsuru_repo`.

The next time we upgrade the packages in our PPA we'll need to think about
how we roll it out to existing environments. Whether we change all of the
playbooks to `state=latest` (which is a pain to variabilise because it's a
separate attribute from the version which we currently expose) or pin the
specific version (which is also a pain because we'd have to time the changes
perfectly). It feels like it's out of scope for now though.