Use wal-e playbook with deletion of old backups
We want to delete older backups from WAL-E, for which we need to setup
a cronjob which will delete the old backups. The [ansible role for
wal-e](https://github.com/alphagov/ansible-playbook-wal-e/pull/5)
implements this feature.

Default values are OK for the time being.

As [commented in the wal-e role
PR](https://github.com/alphagov/ansible-playbook-wal-e/pull/5#issuecomment-126633032) we will
disable the job on GCE until [the bug in WAL-E deleting
keys](https://github.com/wal-e/wal-e/issues/192) is fixed.