Rename var hipache_port to router_port
To reflect the fact that it will be used for both Hipache and Vulcand. We
can move it to `group_vars` at the same time, to match `docker_port` and
`redis_port`, which makes it slightly clearer that the variable name is
different from the one being passed to the actual role.