Optionally install vulcand on the router nodes
Instead of Hipache, when the `vulcand` feature flag is enabled. Version is
the latest stable. Port is the same as Hipache. It's using the single etcd
endpoint which is running on the DB host.

Per the comment in `requirements.yml`, I haven't pinned the version of the
playbook because it may be under active development while we're testing and
it'll be easier to not make two stage releases. We should version it if we
continue to use it in the future.

NB: This isn't intended to convert an existing environment from Hipache to
Vulcand. The ports will clash. We only intend to create new environments,
for testing, with this feature flag.