tsuru_db: Trim vulcand etcd condition
Remove `and vulcand` from `when: vulcand is defined and vulcand`. We don't want to test for specific contents of vulcand variable, just whether it is defined or not. This way we can use the value of variable to carry some other useful information, like vulcand version, or some specific setting.