Optionally configure Tsuru API for vulcand
When the `vulcand` feature flag is enabled. The playbook version is being
bumped at the same time (with backwards incompat changes) so that we can
configure the router. I've arbitrarily chosen/set the API port, because we
need to pass the same value to both vulcand and Tsuru API.

I'm going to create tech debt stories for the following:

- the Tsuru API is only using the first router host for Vulcand's API. If we
  use vulcand beyond testing then we should fix this single point of
  failure
- the variable `hipache_host_external_lb` and the DNS hostname that is
  reference should be renamed from "hipache" to "router", so that it's
  agnostic of the actual router we use