Bump tsuru-api version to add queue config
Bump the tsuru version of playbook to include configuration
that setup queue:mongo-url and queue:mongo-database config.

The snapshot version of tsuru requires this configuration to be set, so we
need to populate these variables in order to test vulcand as router.

The change is backwards compatible and can work with the stable version
of tsuru 0.11.2.