Nginx SSL proxy sets Origin header to upstream url
`Origin` header is used to protect againt unauthorized cross-origin use
of a WebSocket server. Tsuru API will not authorise the request if it
does not match with the address and port that is bind to `(localhost:8080)`,
so Nginx proxy will break this.

We override it to the upstream address, which will disable this
security measure. So far this is OK as the tsuru API has no risk of
cross origin attacks.