Rename tsr to tsurud when using snapshots
The `tsr` binary has been renamed to `tsurud` (tsuru/tsuru#1192) in the
snapshot releases which prevents this command from running when we're using
our `vulcand` feature flag. We'll need to update this again when the stable
release is out.

The service is unaffected because the package ships the binary and init
script at the same time.