Generate and set vulcand seal key
We need to [setup a `sealkey` in vulcand](https://vulcand.io/proxy.html#secret-storage)
to store certificate in the etcd engine.

In order to ensure that we generate a unique seal per environment, we
generate and store a new seal if the file `/etc/vulcand.sealkey`.

The seal [is a hexadecimal representation of 32 random bytes]
(https://github.com/mailgun/vulcand/blob/e6f9b2037c582009c999ed3129cbdc94f36da545/vctl/command/secret.go#L17-L24),
and we use bash dark magic to generate it without having to use the
`vctl` command, which is installed by the role and not available before
executing the role.