remove legacy ubuntu docker node configuration
Ubuntu docker nodes are being replaced with coreOS. We don't need to configure these hosts anymore. Part of this config was legacy already: The 2 actions checking if node was already a part of pool and registering it aftewards were needed long time ago for upgrade of existing clusters which were running on older tsuru version which didn't require explicit node allocation to a pool to a newer version which did require explicit allocation. All newly built clusters have nodes explicitly allocated to a pool on 1st registration.