Exclude coreOS nodes from Telegraf install
Ansible uses the local python interpreter to install Telegraf on nodes.

CoreOS does not inlcude Python in the base image and this can not
run this command or indeed be monitored in this way. To enable this story to
progress we've dropped support for Telegraf collection agent on CoreOS nodes.