post-install: push explictly to HEAD:master
git push to `HEAD:master` says that you want to push `HEAD` (latest commit) of the branch (or commit, if detached) you are on to the remote branch called `master`. Previously we were always pushing the local `master` branch to the remote `master` branch, which wasn't having any effect when we specified a commit reference for the local repository and it checked out "detached HEAD" without updating `master`.