Push unambiguous reference branch in gandalf
When deploying to the remote repository in gandalf, we might hit this error if the remote master branch does not exist. 

```
# git push git@ci-gandalf.tsuru2.paas.alphagov.co.uk:elasticsearchapi.git HEAD:master
error: unable to push to unqualified destination: master
The destination refspec neither matches an existing ref on the remote nor
begins with refs/, and we are unable to guess a prefix based on the source ref.
error: failed to push some refs to 'git@ci-gandalf.tsuru2.paas.alphagov.co.uk:elasticsearchapi.git'
```

This happens if you are using a detach repository (as we do to be able to use pined versions of code in b49b59588fb77a1587113ebb7a39dbeff06645a8). To ensure that the push works we need to use an unambiguous reference.