tsuru_api: remove zombie nodes from default pool
Remove nodes which are currently in the default pool, but are not "ours" - the IPs don't belong to any of the current docker servers. Most common way this situation is created is when we update coreOS nodes configuration (via modifying user-metadata), which re-creates these nodes with new IPs. The new nodes are registered, but the old IPs were staying present in the pool. This commit cleans up all nodes that should not be present in the pool.