Use URLs instead of IPs for Docker nodes
We weren't comparing existing and expected nodes like-for-like. This meant
that if you added a non-existent node with a non-standard port (as described
in the pull request) then it would give an error when trying to remove it:

    stderr: Error: node with address "http://1.2.3.4:4243" not found in cluster

Because the port was taken from the `docker_port` variable rather than the
output of `docker-node-list`. Fix this by getting/constructing URLs before
compare/add/remove.

In order to extract the right field out of `docker-node-list` I'm assuming
that all nodes use HTTP or HTTPS I think that should be OK. We no longer
need to pass it through `sort|uniq` because we are discarding the other
fields where IPs may have been duplicated.