Populate tags with ip and jobname for telegraf
Add the `host_ip` and `job` based on the private IP of the machine and
the vm tag name provided by terraform.

Depending of the platform, we get the vm name from different variables
provided by the Dynamic Inventory Script.

For AWS `ec2_tag_Name` is used, for GCE, `gce_name`