Workaround issue with output of regex in ansible
We found a weird behaviour in ansible. If one passes the output of the
`replace_regex` directly as variables of the role, it might get expanded
as a weird character in ansible (Ctrl+A).

The solution was create a temporary `fact` and pass that one in the
variables of the role.